+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 - Langchain::LLM::Ollama now uses `llama3` by default
+- Langchain::LLM::Anthropic#complete() now uses `claude-2.1` by default
 
 ## [0.12.0] - 2024-04-22
 - [BREAKING] Rename `dimension` parameter to `dimensions` everywhere